various small changes (check CHANGLOG.md)

+++ commands/devTools/ping.js
@@ -30,7 +30,7 @@ module.exports = {
 				const reaction = collected.first();
 				if (reaction.emoji.name === '🗑️') {
 					await m.delete();
-					await message.delete();
+					if (!message.deleted) await message.delete();
 				}
 			})
 			.catch(async collected => {
