+++ commands/help.js
@@ -91,7 +91,7 @@ module.exports = {
 				const reaction = collected.first();
 				if (reaction.emoji.name === '🗑️') {
 					await embedMessage.delete();
-					await message.delete();
+					if (!message.deleted) await message.delete();
 				}
 			})
 			.catch(async collected => {
